# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MOTHERSHIP Multi-Agent Governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/{symbol}:
    get:
      operationId: getMarketSnapshot
      tags: [market]
      summary: Get a one-shot market snapshot for a perp symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketSnapshot"
        "400":
          description: Invalid symbol format or unknown perp symbol
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "502":
          description: Upstream Hyperliquid Info API unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /market/{symbol}/candles:
    get:
      operationId: getMarketCandles
      tags: [market]
      summary: Get historical OHLCV candles for a perp symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: false
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h, 1d]
            default: 1m
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 500
      responses:
        "200":
          description: Candle batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleBatch"
        "400":
          description: Invalid symbol, interval, or limit
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "502":
          description: Upstream Hyperliquid Info API unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /market/{symbol}/stream:
    get:
      operationId: streamMarketSnapshot
      tags: [market]
      summary: Live SSE stream of market snapshots (~2s cadence)
      description: |
        Server-Sent Events stream. Consume from the browser with `EventSource`,
        not via the generated React Query hook (which models this as a single
        fetch and will not deliver subsequent ticks). Each `message` event
        carries a JSON `MarketSnapshot`. Upstream failures are emitted as
        `event: upstream_error` with `{ message: string }`.
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of MarketSnapshot payloads
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Invalid symbol format or unknown perp symbol
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "502":
          description: Upstream Hyperliquid Info API unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  schemas:
...
    Candle:
      type: object
      properties:
        time:
          type: integer
          description: UTC seconds (open time)
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required: [time, open, high, low, close, volume]
    CandleBatch:
      type: object
      properties:
        symbol:
          type: string
        interval:
          type: string
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
      required: [symbol, interval, candles]
    MarketSnapshot:
      type: object
      properties:
        symbol:
          type: string
        markPrice:
          type: number
        oraclePrice:
          type: number
        midPrice:
          type: number
          nullable: true
        bid:
          type: number
          nullable: true
        ask:
          type: number
          nullable: true
        bidSize:
          type: number
          nullable: true
        askSize:
          type: number
          nullable: true
        spread:
          type: number
          nullable: true
        spreadBps:
          type: number
          nullable: true
        fundingRate:
          type: number
        openInterest:
          type: number
        prevDayPrice:
          type: number
        dayVolumeUsd:
          type: number
        changePct24h:
          type: number
        timestamp:
          type: number
          description: Unix epoch milliseconds
      required:
        - symbol
        - markPrice
        - oraclePrice
        - fundingRate
        - openInterest
        - prevDayPrice
        - dayVolumeUsd
        - changePct24h
        - timestamp
...